@@ -2,32 +2,30 @@ source 'https://rubygems.org'
 
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.7'
+gem 'rails', '5.0.0.1'
 # Use mysql as the database for Active Record
 gem 'mysql2', '>= 0.3.13', '< 0.5'
+# Use Puma as the app server
+#gem 'puma', '~> 3.0'
 # Use SCSS for stylesheets
 gem 'sass-rails', '~> 5.0'
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
 # Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 4.1.0'
+gem 'coffee-rails', '~> 4.2'
 # See https://github.com/rails/execjs#readme for more supported runtimes
 gem 'therubyracer', platforms: :ruby
 
 # Use jquery as the JavaScript library
 #gem 'jquery-rails'
-# Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
-#gem 'turbolinks'
+# Turbolinks makes navigating your web application faster. Read more: https://github.com/turbolinks/turbolinks
+#gem 'turbolinks', '~> 5'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.0'
-# bundle exec rake doc:rails generates the API under doc/api.
-gem 'sdoc', '~> 0.4.0', group: :doc
-
+gem 'jbuilder', '~> 2.5'
+# Use Redis adapter to run Action Cable in production
+# gem 'redis', '~> 3.0'
 # Use ActiveModel has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
-
-# Use Unicorn as the app server
-# gem 'unicorn'
 
 # Use Capistrano for deployment
 # gem 'capistrano-rails', group: :development
@@ -35,41 +33,43 @@ gem 'sdoc', '~> 0.4.0', group: :doc
 gem 'activerecord-import', '~> 0.15.0'
 gem 'activerecord-session_store', '~> 1.0.0'
 gem 'active_hash', '~> 1.4.1'
-gem 'enumerize', '~> 1.1.1'
+gem 'enumerize', '~> 2.0.0'
 gem 'ruby-ldap', '~> 0.9.11'
-gem 'will_paginate', '~> 3.1.0'
-gem 'jpmobile', '~> 4.2.0'
-gem 'rmagick', '~> 2.15.4'
+gem 'will_paginate', '~> 3.1.3'
+gem 'jpmobile', '~> 5.0.0'
+gem 'rmagick', '~> 2.16.0'
 gem 'nokogiri', '~> 1.6.8'
-gem 'nokogumbo', '~> 1.4.7'
+gem 'nokogumbo', '~> 1.4.9'
 gem 'hpricot', '~> 0.8.6'
 gem 'premailer', '~> 1.8.7'
-gem 'sanitize', '~> 4.0.1'
+gem 'sanitize', '~> 4.3.0'
 gem 'zipruby', '0.3.6'
-gem 'mime-types', '2.6.1'
+gem 'mime-types', '~> 3.1.0'
 gem 'shared-mime-info', '~> 0.2.5'
-gem 'mail', '2.6.4.rc2'
-gem 'mail-iso-2022-jp', '2.0.5.rc1'
+gem 'mail', '2.6.4'
+gem 'mail-iso-2022-jp', '2.0.5'
 gem 'charlock_holmes', '~> 0.7.3'
 gem 'delayed_job_active_record', '~> 4.1.1'
-gem 'daemons', '~> 1.2.3'
+gem 'daemons', '~> 1.2.4'
 
 gem 'dynamic_form', '~> 1.1.4'
 
-gem 'bower-rails', '~> 0.10.0'
-gem 'whenever', '~> 0.9.4', require: false
+gem 'bower-rails', '~> 0.11.0'
+gem 'whenever', '~> 0.9.7', require: false
 
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
-  gem 'byebug'
+  gem 'byebug', platform: :mri
 end
 
 group :development do
-  # Access an IRB console on exception pages or by using <%= console %> in views
-  gem 'web-console', '~> 2.0'
-
+  # Access an IRB console on exception pages or by using <%= console %> anywhere in the code.
+  gem 'web-console'
+  gem 'listen', '~> 3.0.5'
   # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
   gem 'spring'
-
-  gem 'quiet_assets', '~> 1.1.0'
+  gem 'spring-watcher-listen', '~> 2.0.0'
 end
+
+# Windows does not include zoneinfo files, so bundle the tzinfo-data gem
+gem 'tzinfo-data', platforms: [:mingw, :mswin, :x64_mingw, :jruby]
