@@ -5,47 +5,46 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
 
   before_action :handle_mailto_scheme, if: -> { params[:src] == 'mailto' && params[:uri] }
   before_action :check_user_email, only: [:new, :create, :edit, :update, :answer, :forward]
-  before_action :check_posted_uids, only: [:move, :delete, :seen, :unseen, :register_spam]
+  before_action :check_posted_uids, only: [:move, :delete, :seen, :unseen, :junk]
 
   before_action :set_conf, only: [:index, :show, :move]
   before_action :set_address_histories, only: [:index, :show, :move]
   before_action :set_mail_form_size
 
-  before_action :set_mailboxes, only: [:index, :show, :move, :empty]
-  after_action :reload_mailboxes, only: [:destroy, :delete, :empty, :move, :seen, :unseen, :star, :register_spam]
-
   before_action :set_quota, only: [:index]
-  after_action :reload_quota, only: [:destroy, :delete, :empty, :move, :seen, :unseen, :star, :register_spam]
-
   before_action :set_conditions_from_params, only: [:index, :show]
   before_action :set_sort_from_params, only: [:index, :show, :move]
+
+  after_action :reload_mailboxes, only: [:destroy, :delete, :empty, :move, :seen, :unseen, :star, :junk]
+  after_action :reload_quota, only: [:destroy, :delete, :empty, :move, :seen, :unseen, :star, :junk]
 
   def pre_dispatch
     return redirect_to action: :index, mailbox: params[:mailbox] if params[:reset]
 
-    @mailbox = Webmail::Mailbox.load_mailbox(params[:mailbox] || 'INBOX')
+    @mailboxes = Webmail::Mailbox.load_mailboxes(params[:reload].present? ? :all : nil)
+    @mailbox = @mailboxes.detect { |mailbox| mailbox.name == params[:mailbox] }
     return http_error(404) unless @mailbox
   end
 
   def index
-    last_uid, recent, delayed = Webmail::Filter.apply_recents
+    last_uid, recent, delayed = Webmail::Filter.apply_recents(@mailboxes.detect(&:inbox?))
     if recent
       reload_mailboxes
       reload_quota
     end
-    if @mailbox.name == 'INBOX'
+    if @mailbox.inbox?
       @conditions += ['UID', "1:#{last_uid}"]
     end
     if delayed > 0
       flash.now[:error] = "新着メールのフィルター処理件数が規定値を超えたため、残り#{delayed}件はバックグラウンドで実行します。完了までに時間がかかる場合があります。"
     end
 
-    @items = Webmail::Mail.paginate(select: @mailbox.name, conditions: @conditions,
-      sort: @sort, page: params[:page], limit: @conf.mails_per_page, starred: params[:sort_starred])
+    @items = @mailbox.paginate_mails(conditions: @conditions, sort: @sort,
+      page: params[:page], limit: @conf.mails_per_page, starred: params[:sort_starred])
   end
 
   def show
-    @item = Webmail::Mail.find_by_uid(params[:id], select: @mailbox.name, conditions: ['UNDELETED'])
+    @item = @mailbox.find_mail_by_uid(params[:id], conditions: ['UNDELETED'])
     return error_auth unless @item
 
     if params[:show_html_image]
@@ -57,22 +56,13 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
 
     Core.title += " - #{@item.subject} - #{Core.current_user.email}"
 
-    if from = Email.parse(@item.friendly_from_addr)
-      @from_addr = CGI.escapeHTML(from.address)
-      @from_name = ::NKF::nkf('-wx --cp932', from.name).gsub(/\0/, "") if from.name rescue nil
-      @from_name = @from_name || @from_addr
-    end
-
     if @item.unseen?
-      mailbox_uids = get_mailbox_uids(@mailbox, @item.uid)
-      mailbox_uids.each do |mailbox, uids|
-        Webmail::Mail.seen_all(mailbox, uids)
-      end
+      @mailbox.flag_mails(@item.uid, [:Seen])
       reload_mailboxes
 
       @item.seen!
 
-      if @item.mdn_request_mode == :auto && !@item.notified? && !@mailbox.draft_box? && !@mailbox.sent_box?
+      if @item.mdn_request_mode == :auto && !@item.mdn_sent? && !@mailbox.draft_box? && !@mailbox.sent_box?
         begin
           send_mdn_message(@item.mdn_request_mode)
         rescue => e
@@ -81,14 +71,14 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
       end
     end
 
-    @pagination = Webmail::Mail.paginate_uid(params[:id],
-      select: @mailbox.name, conditions: @conditions, sort: @sort, starred: params[:sort_starred])
+    @pagination = @mailbox.paginate_mail_by_uid(params[:id], conditions: @conditions, sort: @sort,
+      starred: params[:sort_starred])
 
     _show @item
   end
 
   def download
-    @item = Webmail::Mail.find_by_uid(params[:id], select: @mailbox.name, conditions: ['UNDELETED'])
+    @item = @mailbox.find_mail_by_uid(params[:id], conditions: ['UNDELETED'])
     return error_auth unless @item
 
     case
@@ -139,7 +129,7 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
   end
 
   def edit
-    @ref = Webmail::Mail.find_by_uid(params[:id], select: @mailbox.name, conditions: ['UNDELETED'])
+    @ref = @mailbox.find_mail_by_uid(params[:id], conditions: ['UNDELETED'])
     return http_error(404) unless @ref
 
     @item = Webmail::Mail.new(item_params)
@@ -150,25 +140,19 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
   end
 
   def update
-    @ref = Webmail::Mail.find_by_uid(params[:id], select: @mailbox.name, conditions: ['UNDELETED'])
+    @ref = @mailbox.find_mail_by_uid(params[:id], conditions: ['UNDELETED'])
     return http_error(404) unless @ref
 
     @item = Webmail::Mail.new(item_params)
     send_message(@item, @ref) do
       if !params[:remain_draft] && @ref.draft?
-        mailbox_uids = get_mailbox_uids(@mailbox, @ref.uid)
-        mailbox_uids.each do |mailbox, uids|
-          num = Webmail::Mail.delete_all(mailbox, uids, true)
-          if num > 0
-            Webmail::MailNode.delete_nodes(mailbox, uids)
-          end
-        end
+        @mailbox.delete_mails(@ref.uid)
       end
     end
   end
 
   def answer
-    @ref = Webmail::Mail.find_by_uid(params[:id], select: @mailbox.name, conditions: ['UNDELETED'])
+    @ref = @mailbox.find_mail_by_uid(params[:id], conditions: ['UNDELETED'])
     return http_error(404) unless @ref
 
     @item = Webmail::Mail.new(item_params)
@@ -176,11 +160,7 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
 
     if request.post?
       return send_message(@item, @ref) do
-        mailbox_uids = get_mailbox_uids(@mailbox, @ref.uid)
-        mailbox_uids.each do |mailbox, uids|
-          Core.imap.select(mailbox)
-          Core.imap.uid_store(uids, "+FLAGS", [:Answered])
-        end
+        @mailbox.flag_mails(@ref.uid, [:Answered])
       end
     end
 
@@ -196,18 +176,14 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
   end
 
   def forward
-    @ref = Webmail::Mail.find_by_uid(params[:id], select: @mailbox.name, conditions: ['UNDELETED'])
+    @ref = @mailbox.find_mail_by_uid(params[:id], conditions: ['UNDELETED'])
     return http_error(404) unless @ref
 
     @item = Webmail::Mail.new(item_params)
 
     if request.post?
       return send_message(@item, @ref) do
-        mailbox_uids = get_mailbox_uids(@mailbox, @ref.uid)
-        mailbox_uids.each do |mailbox, uids|
-          Core.imap.select(mailbox)
-          Core.imap.uid_store(uids, "+FLAGS", "$Forwarded")
-        end
+        @mailbox.flag_mails(@ref.uid, ['$Forwarded'])
       end
     end
 
@@ -254,7 +230,7 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
 
     begin
       mail = item.prepare_mail(request)
-      mail.delivery_method(:smtp, ActionMailer::Base.smtp_settings)
+      mail.delivery_method(:smtp, Webmail::Mail.smtp_settings(Core.current_user))
       mail.deliver
     rescue => e
       flash.now[:error] = "メールの送信に失敗しました。（#{e}）"
@@ -273,7 +249,8 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
     ## save to 'Sent'
     begin
       mail.header[:bcc].include_in_headers = true
-      Timeout.timeout(60) { Core.imap.append('Sent', mail.to_s, [:Seen], Time.now) }
+      sent = @mailboxes.detect(&:use_as_sent?)
+      Timeout.timeout(60) { Core.imap.append(sent.name, mail.to_s, [:Seen], Time.now) }
     rescue => e
       flash[:error] = "メールは送信できましたが、送信トレイへの保存に失敗しました。（#{e}）"
     end
@@ -291,7 +268,8 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
       mail = item.prepare_mail(request)
       mail.header[:bcc].include_in_headers = true
       flags = [:Seen, :Draft].tap { |a| a << :Flagged if ref && ref.starred? }
-      Timeout.timeout(60) { Core.imap.append('Drafts', mail.to_s, flags, Time.now) }
+      draft = @mailboxes.detect(&:use_as_drafts?)
+      Timeout.timeout(60) { Core.imap.append(draft.name, mail.to_s, flags, Time.now) }
     rescue => e
       flash.now[:error] = "下書き保存に失敗しました。（#{e}）"
       respond_to do |format|
@@ -314,41 +292,14 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
   end
 
   def destroy
-    @item = Webmail::Mail.find_by_uid(params[:id], select: @mailbox.name, conditions: ['UNDELETED'], fetch: ['FLAGS'])
+    @item = @mailbox.find_mail_by_uid(params[:id], conditions: ['UNDELETED'], fetch: ['FLAGS'])
     return error_auth unless @item
 
-    changed_num = 0
-    changed_mailbox_uids = {}
+    trash = @mailboxes.detect(&:use_as_trash?)
+    num = @mailbox.trash_mails(trash.name, @item.uid)
 
-    mailbox_uids = get_mailbox_uids(@mailbox, @item.uid)
-    mailbox_uids.each do |mailbox, uids|
-      num = Webmail::Mail.delete_all(mailbox, uids)
-      if num > 0
-        Webmail::MailNode.delete_nodes(mailbox, uids)
-        changed_mailbox_uids['Trash'] = [:all]
-      end
-      changed_num += num if mailbox !~ /^(Star)$/
-    end
-
-    if changed_num > 0
-      reload_starred_mails(changed_mailbox_uids) if @item.starred?
-
-      flash[:notice] = 'メールを削除しました。' unless new_window?
-      respond_to do |format|
-        format.html do
-          redirect_to action: new_window? ? :close : :index
-        end
-        format.xml  { head :ok }
-      end
-    else
-      flash[:error] = 'メールの削除に失敗しました。'
-      respond_to do |format|
-        format.html do
-          redirect_to action: new_window? ? :close : :index
-        end
-        format.xml  { render :xml => @item.errors, :status => :unprocessable_entity }
-      end
-    end
+    flash[:notice] = num > 0 ? 'メールを削除しました。' : 'メールの削除に失敗しました。' unless new_window?
+    redirect_to action: new_window? ? :close : :index
   end
 
   def move
@@ -356,45 +307,19 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
     return http_error if uids.blank?
 
     if !params[:item][:mailbox]
-      @items = Webmail::Mail.find(select: @mailbox.name, conditions: ['UID', uids] + ['UNDELETED'], sort: @sort)
+      @items = @mailbox.find_mails(conditions: ['UID', uids] + ['UNDELETED'], sort: @sort)
       return render template: 'webmail/admin/mails/move'
     end
 
-    changed_num = 0
-    changed_mailbox_uids = {}
-    include_starred_uid = Webmail::Mail.include_starred_uid?(@mailbox.name, uids)
-
-    mailbox_uids = get_mailbox_uids(@mailbox, uids)
-    mailbox_uids.each do |mailbox, uids|
-      if mailbox =~ /^(Star)$/
-        num = 0
-        if params[:copy].blank? #move
-          num = Webmail::Mail.delete_all(mailbox, uids, true)
-          Webmail::MailNode.delete_nodes(mailbox, uids) if num > 0
-        end
+    num =
+      if params[:copy].present?
+        @mailbox.copy_mails(params[:item][:mailbox], uids)
       else
-        if params[:copy].blank? #move
-          num = Webmail::Mail.move_all(mailbox, params[:item][:mailbox], uids)
-          Webmail::MailNode.delete_nodes(mailbox, uids) if num > 0
-        else
-          num = Webmail::Mail.copy_all(mailbox, params[:item][:mailbox], uids)
-        end
-        changed_num += num
+        @mailbox.move_mails(params[:item][:mailbox], uids)
       end
-      if num > 0
-        changed_mailbox_uids[params[:item][:mailbox]] = [:all]
-      end
-    end
-
-    reload_starred_mails(changed_mailbox_uids) if include_starred_uid
-
-#    num = 0
-#    Webmail::Mail.find(select: @mailbox.name, conditions: cond).each do |item|
-#      num += 1 if item.move(params[:item][:mailbox])
-#    end
 
     label = params[:copy].blank? ? '移動' : 'コピー'
-    flash[:notice] = "#{changed_num}件のメールを#{label}しました。" unless new_window?
+    flash[:notice] = "#{num}件のメールを#{label}しました。" unless new_window?
     redirect_to action: new_window? ? :close : :index
   end
 
@@ -402,27 +327,10 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
     uids = params[:item][:ids].keys.map(&:to_i).select(&:positive?)
     return http_error if uids.blank?
 
-    changed_num = 0
-    changed_mailbox_uids = {}
-    include_starred_uid = Webmail::Mail.include_starred_uid?(@mailbox.name, uids)
+    trash = @mailboxes.detect(&:use_as_trash?)
+    num = @mailbox.trash_mails(trash.name, uids)
 
-    mailbox_uids = get_mailbox_uids(@mailbox, uids)
-    mailbox_uids.each do |mailbox, uids|
-      num = Webmail::Mail.delete_all(mailbox, uids)
-      if num > 0
-        Webmail::MailNode.delete_nodes(mailbox, uids)
-        changed_mailbox_uids['Trash'] = [:all]
-      end
-      changed_num += num if mailbox !~ /^(Star)$/
-    end
-
-    reload_starred_mails(changed_mailbox_uids) if include_starred_uid
-
-#    Webmail::Mail.find(select: @mailbox.name, conditions: cond).each do |item|
-#      num += 1 if item.destroy
-#    end
-
-    flash[:notice] = "#{changed_num}件のメールを削除しました。"
+    flash[:notice] = "#{num}件のメールを削除しました。"
     redirect_to action: :index
   end
 
@@ -430,14 +338,9 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
     uids = params[:item][:ids].keys.map(&:to_i).select(&:positive?)
     return http_error if uids.blank?
 
-    changed_num = 0
-    mailbox_uids = get_mailbox_uids(@mailbox, uids)
-    mailbox_uids.each do |mailbox, uids|
-      num = Webmail::Mail.seen_all(mailbox, uids)
-      changed_num += num if mailbox !~ /^(Star)$/
-    end
+    num = @mailbox.flag_mails(uids, [:Seen])
 
-    flash[:notice] = "#{changed_num}件のメールを既読にしました。"
+    flash[:notice] = "#{num}件のメールを既読にしました。"
     redirect_to action: :index
   end
 
@@ -445,100 +348,57 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
     uids = params[:item][:ids].keys.map(&:to_i).select(&:positive?)
     return http_error if uids.blank?
 
-    changed_num = 0
-    mailbox_uids = get_mailbox_uids(@mailbox, uids)
-    mailbox_uids.each do |mailbox, uids|
-      num = Webmail::Mail.unseen_all(mailbox, uids)
-      changed_num += num if mailbox !~ /^(Star)$/
-    end
+    num = @mailbox.unflag_mails(uids, [:Seen])
 
-    flash[:notice] = "#{changed_num}件のメールを未読にしました。"
+    flash[:notice] = "#{num}件のメールを未読にしました。"
     redirect_to action: :index
   end
 
-  def register_spam
+  def junk
     uids = params[:item][:ids].keys.map(&:to_i).select(&:positive?)
     return http_error if uids.blank?
 
-    items = Webmail::Mail.find(select: @mailbox.name, conditions: ['UID', uids] + ['UNDELETED'])
+    items = @mailbox.find_mails(conditions: ['UID', uids] + ['UNDELETED'])
     return redirect_to action: :index if items.blank?
 
     Webmail::Filter.register_spams(items)
 
-    changed_num = 0
-    changed_mailbox_uids = {}
-    include_starred_uid = Webmail::Mail.include_starred_uid?(@mailbox.name, uids)
-
-    mailbox_uids = get_mailbox_uids(@mailbox, uids)
-    mailbox_uids.each do |mailbox, uids|
-      num = Webmail::Mail.delete_all(mailbox, uids)
-      if num > 0
-        Webmail::MailNode.delete_nodes(mailbox, uids)
-        changed_mailbox_uids['Trash'] = [:all]
-      end
-      changed_num += num if mailbox !~ /^(Star)$/
-    end
-
-    reload_starred_mails(changed_mailbox_uids) if include_starred_uid
+    junk = @mailboxes.detect(&:use_as_junk?)
+    @mailbox.move_mails(junk.name, uids)
 
     flash[:notice] = "#{items.count}件のメールを迷惑メールに登録しました。"
     redirect_to action: :index
   end
 
   def empty
-    changed_mailbox_uids = {}
-
-    @mailboxes.reverse.each do |box|
-      if box.trash_box?(:children)
-        begin
-          Webmail::MailNode.delete_nodes(box.name)
-          Core.imap.delete(box.name)
-
-          uids = Webmail::MailNode.find_ref_nodes(box.name).map{|x| x.uid}
-          num = Webmail::Mail.delete_all('Star', uids)
-          if num > 0
-            Webmail::MailNode.delete_ref_nodes(box.name)
-            changed_mailbox_uids[box.name] = [:all]
-          end
-        rescue => e
-        end
+    @mailboxes.reverse.each do |mailbox|
+      if mailbox.trash_box?(:children)
+        mailbox.delete_mailbox
       end
     end
 
-    Core.imap.select(@mailbox.name)
-    uids = Core.imap.uid_search(['UNDELETED'], 'utf-8')
+    @mailbox.empty_mails
 
-    mailbox_uids = get_mailbox_uids(@mailbox, uids)
-    mailbox_uids.each do |mailbox, uids|
-      num = Webmail::Mail.delete_all(mailbox, uids, true)
-      if num > 0
-        Webmail::MailNode.delete_nodes(mailbox, uids)
-        changed_mailbox_uids[mailbox] = [:all]
-      end
-    end
-
-    reload_starred_mails(changed_mailbox_uids)
-
-    flash[:notice] = "ごみ箱を空にしました。"
+    flash[:notice] = 'ごみ箱を空にしました。'
     respond_to do |format|
       format.html { redirect_to action: :index }
       format.xml  { head :ok }
     end
   end
-  
+
   def send_mdn
-    @item = Webmail::Mail.find_by_uid(params[:id], select: @mailbox.name, conditions: ['UNDELETED'])
+    @item = @mailbox.find_mail_by_uid(params[:id], conditions: ['UNDELETED'])
     return error_auth unless @item && @item.has_disposition_notification_to?
 
     if request.xhr?
       send_mdn_message(:manual)
-      return render text: ''
+      return render plain: ''
     else
       begin
         send_mdn_message(:manual)
-        flash[:notice] = "開封確認メールを送信しました。"
+        flash[:notice] = '開封確認メールを送信しました。'
       rescue => e
-        flash[:notice] = "開封確認メールの送信に失敗しました。"
+        flash[:notice] = '開封確認メールの送信に失敗しました。'
       end
       return redirect_to action: :show
     end
@@ -552,63 +412,41 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
   end
 
   def star
-    @item = Webmail::Mail.find_by_uid(params[:id], select: @mailbox.name, conditions: ['UNDELETED'], fetch: ['FLAGS'])
+    @item = @mailbox.find_mail_by_uid(params[:id], conditions: ['UNDELETED'], fetch: ['FLAGS'])
     return error_auth unless @item
 
-    starred = @item.starred?
-    changed_mailbox_uids = {}
-
-    mailbox_uids = get_mailbox_uids(@mailbox, @item.uid)
-    mailbox_uids.each do |mailbox, uids|
-      if starred
-        num = Webmail::Mail.unstar_all(mailbox, uids)
-      else
-        num = Webmail::Mail.star_all(mailbox, uids)
-      end
-      if num > 0
-        changed_mailbox_uids[mailbox] = uids
-      end
+    if @item.starred?
+      @mailbox.unflag_mails(@item.uid, [:Flagged])
+    else
+      @mailbox.flag_mails(@item.uid, [:Flagged])
     end
 
-    reload_starred_mails(changed_mailbox_uids)
-
     if request.mobile?
-      if params[:from] == 'list' || @mailbox.star_box?
+      if params[:from] == 'list' || @mailbox.flagged_box?
         redirect_to action: :index, id: params[:id], mobile: :list
       else
         redirect_to action: :show, id: params[:id]
       end
     else
-      render text: "OK"
+      render plain: 'OK'
     end
   end
 
   def label
-    @item = Webmail::Mail.find_by_uid(params[:id], select: @mailbox.name, conditions: ['UNDELETED'], fetch: ['FLAGS'])
+    @item = @mailbox.find_mail_by_uid(params[:id], conditions: ['UNDELETED'], fetch: ['FLAGS'])
     return error_auth unless @item
 
     @label_confs = Webmail::Setting.load_label_confs
 
     label_id = params[:label].to_i
-    labeled = @item.labeled?(label_id)
+    labels = label_id == 0 ? (1..9).map { |id| "$label#{id}" } : ["$label#{label_id}"]
 
-    mailbox_uids = get_mailbox_uids(@mailbox, @item.uid)
-    mailbox_uids.each do |mailbox, uids|
-      if label_id == 0
-        Webmail::Mail.unlabel_all(mailbox, uids)
-      elsif labeled
-        Webmail::Mail.unlabel_all(mailbox, uids, label_id)
-      else
-        Webmail::Mail.label_all(mailbox, uids, label_id)
-      end
-    end
-
-    if label_id == 0
-      @item.flags.clear
-    elsif labeled
-      @item.flags.delete("$label#{label_id}")
+    if label_id == 0 || @item.labeled?(label_id)
+      @mailbox.unflag_mails(@item.uid, labels)
+      @item.flags -= labels
     else
-      @item.flags << "$label#{label_id}"
+      @mailbox.flag_mails(@item.uid, labels)
+      @item.flags += labels
     end
 
     render layout: false
@@ -634,7 +472,7 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
   def keep_params(options = {})
     if options[:mailbox].blank? && options[:controller].blank?
       keeps = params.slice(:page, :search, :s_keyword, :s_column, :s_status, :s_label,
-        :sort_key, :sort_order, :sort_starred, :new_window)
+        :sort_key, :sort_order, :sort_starred, :new_window).permit!
       options = options.reverse_merge(keeps)
     end
     options
@@ -648,7 +486,7 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
 
   def check_user_email
     if Core.current_user.email.blank?
-      return render text: "メールアドレスが登録されていません。", layout: true
+      return render plain: 'メールアドレスが登録されていません。', layout: true
     end
   end
 
@@ -660,8 +498,11 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
 
   def item_params
     return {} unless params[:item]
-    params.require(:item).permit(:in_to, :in_cc, :in_bcc, :in_subject, :in_body, :in_html_body,
-      :in_format, :in_request_mdn, :tmp_id, :tmp_attachment_ids => [], :in_files => [])
+    params.require(:item).permit(
+      :in_to, :in_cc, :in_bcc, :in_reply_to, :in_subject, :in_body, :in_html_body,
+      :in_format, :in_priority, :in_request_mdn, :in_request_dsn,
+      :tmp_id, :tmp_attachment_ids => [], :in_files => []
+    )
   end
 
   def set_conf
@@ -688,22 +529,16 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
     @mail_form_size ||= Webmail::Setting.user_config_value(:mail_form_size, 'medium')
   end
 
-  def set_mailboxes
-    if !new_window?
-      @mailboxes = Webmail::Mailbox.load_mailboxes(params[:reload].present? ? :all : nil)
-    end
-  end
-
   def reload_mailboxes
     @mailboxes = Webmail::Mailbox.load_mailboxes(:all)
   end
 
   def set_quota
-    @quota = Webmail::Mailbox.load_quota(params[:reload].present?)
+    @quota = Webmail::QuotaRoot.load_quota(params[:reload].present?)
   end
 
   def reload_quota
-    @quota = Webmail::Mailbox.load_quota(true)
+    @quota = Webmail::QuotaRoot.load_quota(true)
   end
 
   def set_conditions_from_params
@@ -713,12 +548,6 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
 
   def set_sort_from_params
     @sort = Webmail::Mail.make_sort_from_params(params)
-  end
-
-  def reload_starred_mails(mailbox_uids = {'INBOX' => :all})
-    Util::Database.lock_by_name(Core.current_user.account) do
-      Webmail::Mailbox.load_starred_mails(mailbox_uids)
-    end
   end
 
   def default_sign
@@ -737,32 +566,11 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
     mdn = Webmail::Mail.new
     mdn.in_from ||= Core.current_user.email_format
     mail = mdn.prepare_mdn(@item, mdn_mode.to_s, request)
-    mail.delivery_method(:smtp, ActionMailer::Base.smtp_settings)
+    mail.delivery_method(:smtp, Webmail::Mail.smtp_settings(Core.current_user))
     mail.deliver
 
-    Core.imap.select(@mailbox.name)
-    Core.imap.uid_store(@item.uid, "+FLAGS", "$Notified")
-    @item.flags << "$Notified"
-  end
-
-  def get_mailbox_uids(mailbox, uids)
-    mailbox_uids = {}
-    uids = [uids] if uids.is_a?(Fixnum)
-    if mailbox.star_box?
-      nodes = Webmail::MailNode.find_nodes(mailbox.name, uids)
-      nodes = nodes.select{|x| x.ref_mailbox && x.ref_uid}.group_by(&:ref_mailbox)
-      nodes.each do |k,v| 
-        mailbox_uids[k] = v.map{|x| x.ref_uid}
-      end
-    else
-      nodes = Webmail::MailNode.find_ref_nodes(mailbox.name, uids)
-      nodes = nodes.group_by(&:mailbox)
-      nodes.each do |k,v| 
-        mailbox_uids[k] = v.map{|x| x.uid}
-      end
-    end
-    mailbox_uids[mailbox.name] = uids
-    mailbox_uids
+    @mailbox.flag_mails(@item.uid, ['$MDNSent'])
+    @item.flags << '$MDNSent'
   end
 
   def rescue_mail_action?
