@@ -1,6 +1,5 @@
-class Webmail::Doc < ActiveRecord::Base
+class Webmail::Doc < ApplicationRecord
   include Sys::Model::Base
-  include Sys::Model::Base::Page
   include Sys::Model::Auth::Manager
 
   belongs_to_active_hash :status, foreign_key: :state, class_name: 'Sys::Base::Status'
@@ -11,5 +10,9 @@ class Webmail::Doc < ActiveRecord::Base
 
   def readable?
     true
+  end
+
+  def states
+    [['公開','public'],['非公開','closed']]
   end
 end
