@@ -1,10 +1,8 @@
-class Webmail::MailNode < ActiveRecord::Base
+class Webmail::MailNode < ApplicationRecord
   include Sys::Model::Base
   include Sys::Model::Auth::Free
 
   validates :user_id, :uid, :mailbox, presence: true
-
-  scope :readable, ->(user = Core.current_user) { where(user_id: user.id) }
 
   def editable?
     Core.current_user.has_auth?(:manager) || user_id == Core.current_user.id
@@ -15,28 +13,8 @@ class Webmail::MailNode < ActiveRecord::Base
   end
 
   class << self
-    def find_nodes(boxname, uids = nil)
-      items = self.where(user_id: Core.current_user.id, mailbox: boxname)
-      items = items.where(uid: uids) if uids
-      items
-    end
-
-    def find_ref_nodes(boxname, uids = nil)
-      items = self.where(user_id: Core.current_user.id, ref_mailbox: boxname)
-      items = items.where(ref_uid: uids) if uids
-      items
-    end
-
-    def delete_nodes(boxname, uids = nil)
-      items = self.where(user_id: Core.current_user.id, mailbox: boxname)
-      items = items.where(uid: uids) if uids
-      items.delete_all
-    end
-
-    def delete_ref_nodes(boxname, uids = nil)
-      items = self.where(user_id: Core.current_user.id, ref_mailbox: boxname)
-      items = items.where(ref_uid: uids) if uids
-      items.delete_all
+    def delete_nodes(boxname, uids)
+      self.where(user_id: Core.current_user.id, mailbox: boxname, uid: uids).delete_all
     end
 
     def delete_caches(batch_size = 10000, sleep_sec = 1)
@@ -46,6 +24,10 @@ class Webmail::MailNode < ActiveRecord::Base
         sleep sleep_sec
         yield sliced_ids if block_given?
       end
+    end
+
+    def delete_expired_caches(ex = Joruri.config.application['webmail.mail_cache_expiration'])
+      self.where("created_at < ?", ex.months.ago).delete_all if ex > 0
     end
   end
 end
