@@ -3,7 +3,7 @@ class Sys::Admin::UsersController < Sys::Controller::Admin::Base
 
   def pre_dispatch
     return error_auth unless Core.user.has_auth?(:manager)
-    return redirect_to(request.env['PATH_INFO']) if params[:reset]
+    return redirect_to action: :index if params[:reset]
   end
 
   def index
