@@ -6,7 +6,7 @@ class Webmail::Admin::FiltersController < Webmail::Controller::Admin::Base
   end
 
   def index
-    @items = Webmail::Filter.readable.where(user_id: Core.user.id).order(:sort_no, :id)
+    @items = Webmail::Filter.where(user_id: Core.user.id).order(:sort_no, :id)
       .paginate(page: params[:page], per_page: params[:limit])
     _index @items
   end
@@ -57,17 +57,7 @@ class Webmail::Admin::FiltersController < Webmail::Controller::Admin::Base
 
     @item.apply
 
-    if @item.applied > 0
-      changed_mailbox_uids = {}
-      case @item.action
-      when 'move'
-        changed_mailbox_uids[@item.mailbox] = [:all]
-      when 'delete'
-        changed_mailbox_uids['Trash'] = [:all]
-      end
-      Webmail::Mailbox.load_starred_mails(changed_mailbox_uids)
-      Webmail::Mailbox.load_mailboxes(:all)
-    end
+    Webmail::Mailbox.load_mailboxes(:all)
 
     flash[:notice] = "#{@item.applied}件のメールに適用しました。"
     flash[:error] = "フィルター処理件数が規定値を超えたため、残り#{@item.delayed}件のメールはバックグラウンドで実行します。完了までに時間がかかる場合があります。" if @item.delayed > 0
@@ -77,11 +67,11 @@ class Webmail::Admin::FiltersController < Webmail::Controller::Admin::Base
   private
 
   def item_params
-    params.require(:item).permit(:name, :state, :sort_no, :conditions_chain, :action, :mailbox,
+    params.require(:item).permit(:name, :state, :sort_no, :conditions_chain, :action, :mailbox_name,
       :conditions_attributes => [:id, :column, :inclusion, :value, :_destroy])
   end
 
   def apply_item_params
-    params.require(:item).permit(:target_mailbox, :include_sub)
+    params.require(:item).permit(:apply_mailbox_name, :include_sub)
   end
 end
