@@ -3,7 +3,7 @@ module Webmail::Mails::Imap
   extend ActiveSupport::Concern
 
   attr_accessor :uid, :mailbox, :flags, :rfc822, :size, :mail
-  attr_accessor :x_mailbox, :x_real_uid
+  attr_accessor :priority, :x_mailbox, :x_real_uid
 
   def imap
     Core.imap
@@ -48,8 +48,8 @@ module Webmail::Mails::Imap
     flags.index(:Flagged)
   end
 
-  def notified?
-    flags.index('$Notified')
+  def mdn_sent?
+    flags.index('$Notified') || flags.index('$MDNSent')
   end
 
   def labeled?(label_id = "")
@@ -70,20 +70,11 @@ module Webmail::Mails::Imap
     @seen_flagged
   end
 
-  def x_mailbox_title
-    Webmail::Mailbox.name_to_title(x_mailbox.to_s).split('.').last
-  end
-
-  def destroy(complete = false)
+  def destroy
     imap.select(mailbox)
-    if mailbox !~ /^Trash(\.|$)/ && !complete
-      num = self.class.move_to('Trash', [uid])
-      num == 1
-    else
-      num = imap.uid_store(uid, '+FLAGS', [:Deleted]).to_a.size
-      imap.expunge
-      num == 1
-    end
+    num = imap.uid_store(uid, '+FLAGS', [:Deleted]).to_a.size
+    imap.expunge
+    num == 1
   end
 
   def move(to_mailbox)
@@ -104,7 +95,7 @@ module Webmail::Mails::Imap
       if sort && imap.capabilities.include?('SORT')
         imap.uid_sort(sort, conditions, 'utf-8')
       else
-        imap.uid_search(conditions, 'utf-8')
+        imap.uid_search(conditions, 'utf-8').reverse
       end
     end
 
@@ -116,6 +107,8 @@ module Webmail::Mails::Imap
       search_uid = imap.uid_search(['UID', uid] + conditions, 'utf-8').first
       return nil unless search_uid
 
+      fetch += ['X-MAILBOX', 'X-REAL-UID'] if select =~ /^virtual\./
+
       msg = imap.uid_fetch(search_uid, fetch).to_a.first
       return nil unless msg
 
@@ -124,6 +117,10 @@ module Webmail::Mails::Imap
       item.uid     = uid
       item.mailbox = select
       item.flags   = msg.attr['FLAGS'] if msg.attr['FLAGS']
+      if select =~ /^virtual\./
+        item.x_mailbox = msg.attr['X-MAILBOX']
+        item.x_real_uid = msg.attr['X-REAL-UID']
+      end
       item
     end
 
@@ -159,7 +156,7 @@ module Webmail::Mails::Imap
       end
     end
 
-    def paginate_uid(uid, select:, conditions: [], sort: [], starred: nil)
+    def paginate_by_uid(uid, select:, conditions: [], sort: [], starred: nil)
       uids =
         if starred == '1'
           find_uids(select: select, conditions: conditions + ['FLAGGED'], sort: sort) +
@@ -214,54 +211,23 @@ module Webmail::Mails::Imap
       res.name == 'OK' ? uids.size : 0
     end
 
-    def delete_all(mailbox, uids, complete = false)
+    def delete_all(mailbox, uids)
       return 0 if uids.blank?
 
       imap.select(mailbox)
-      if mailbox !~ /^Trash(\.|$)/ && mailbox !~ /^Star(\.|$)/ && !complete
-        move_to('Trash', uids)
-      else
-        num = imap.uid_store(uids, "+FLAGS", [:Deleted]).to_a.size
-        imap.expunge
-        num
-      end
+      num = imap.uid_store(uids, '+FLAGS', [:Deleted]).to_a.size
+      imap.expunge
+      num
     end
 
-    def seen_all(mailbox, uids)
+    def flag_all(mailbox, uids, flags)
       imap.select(mailbox)
-      imap.uid_store(uids, "+FLAGS", [:Seen]).to_a.size
+      imap.uid_store(uids, '+FLAGS', flags).to_a.size
     end
 
-    def unseen_all(mailbox, uids)
+    def unflag_all(mailbox, uids, flags)
       imap.select(mailbox)
-      imap.uid_store(uids, "-FLAGS", [:Seen]).to_a.size
-    end
-
-    def star_all(mailbox, uids)
-      imap.select(mailbox)
-      imap.uid_store(uids, "+FLAGS", [:Flagged]).to_a.size
-    end
-
-    def unstar_all(mailbox, uids)
-      imap.select(mailbox)
-      imap.uid_store(uids, "-FLAGS", [:Flagged]).to_a.size
-    end
-
-    def include_starred_uid?(mailbox, uids)
-      imap.select(mailbox)
-      starred_uids = imap.uid_search(['UID', uids, 'UNDELETED', 'FLAGGED'], 'utf-8')
-      uids.inject(false){|result,x| result || starred_uids.include?(x)}
-    end
-
-    def label_all(mailbox, uids, label_id)
-      imap.select(mailbox)
-      imap.uid_store(uids, "+FLAGS", ["$label#{label_id}"]).to_a.size
-    end
-
-    def unlabel_all(mailbox, uids, label_id = nil)
-      labels = label_id ? ["$label#{label_id}"] : (1..9).map { |id| "$label#{id}" }
-      imap.select(mailbox)
-      imap.uid_store(uids, "-FLAGS", labels).to_a.size
+      imap.uid_store(uids, '-FLAGS', flags).to_a.size
     end
 
     def move_to(mailbox, uids)
