@@ -15,37 +15,21 @@ class Webmail::Tool::BatchDelete
     sent_since = Time.parse(self.start_date).strftime("%d-%b-%Y")
     sent_before = (Time.parse(self.end_date) + 1.days).strftime("%d-%b-%Y")
 
-    mailboxes.each do |box|
-      next unless box.batch_deletable_box?
-      next if mailbox_id != 0 && mailbox_id != box.id 
+    mailboxes.each do |mailbox|
+      next unless mailbox.batch_deletable_box?
+      next if mailbox_id != 0 && mailbox_id != mailbox.id 
 
       condition = ['SENTSINCE', sent_since, 'SENTBEFORE', sent_before]
       condition << 'UNFLAGGED' if self.include_starred == '0'
 
-      Core.imap.select(box.name)
+      Core.imap.select(mailbox.name)
       uids = Core.imap.uid_search(condition)
-      num = Core.imap.uid_store(uids, "+FLAGS", [:Deleted]).size rescue 0
-      Core.imap.expunge
-
-      if num > 0
-        Webmail::MailNode.delete_nodes(box.name, uids)
-        changed_mailboxes << box.name
-      end
-
-      delete_num += num
-
-      starred_uids = Webmail::MailNode.find_ref_nodes(box.name, uids).map{|x| x.uid}
-      Core.imap.select('Star')
-      num = Core.imap.uid_store(starred_uids, "+FLAGS", [:Deleted]).size rescue 0
-      Core.imap.expunge
-      if num > 0
-        Webmail::MailNode.delete_ref_nodes(box.name, uids)
-      end
+      delete_num += mailbox.delete_mails(uids)
     end
 
     if delete_num > 0
       Webmail::Mailbox.load_mailboxes(:all)
-      Webmail::Mailbox.load_quota(true)
+      Webmail::QuotaRoot.load_quota(true)
     end
 
     delete_num
